+++ src/components/Upload/ChooseThumbnail.tsx
@@ -9,7 +9,7 @@ import { generateVideoThumbnails } from '@utils/functions/generateVideoThumbnail
 import { getFileFromDataURL } from '@utils/functions/getFileFromDataURL'
 import { getIsNSFW } from '@utils/functions/getIsNSFW'
 import { sanitizeIpfsUrl } from '@utils/functions/sanitizeIpfsUrl'
-import uploadToIPFS3 from '@utils/functions/uploadToIPFS3'
+import uploadToIPFS from '@utils/functions/uploadToIPFS'
 import clsx from 'clsx'
 import * as nsfwjs from 'nsfwjs'
 import { ChangeEvent, FC, useEffect, useState } from 'react'
@@ -39,7 +39,7 @@ const ChooseThumbnail: FC<Props> = ({ label, afterUpload, file }) => {
 
   const uploadThumbnailToIpfs = async (file: File) => {
     setUploadedVideo({ uploadingThumbnail: true })
-    const result: IPFSUploadResult = await uploadToIPFS3(file)
+    const result: IPFSUploadResult = await uploadToIPFS(file)
     setUploadedVideo({ uploadingThumbnail: false })
     afterUpload(result.url, file.type || 'image/jpeg')
     return result
